% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasers-package.R
\docType{package}
\name{lasers-package}
\alias{lasers}
\alias{lasers-package}
\title{lasers: Data for Airplane Laser Incidents as Reported to the FAA}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A dataset of laser incidents as reported by U.S. pilots to the Federal Aviation Admnistration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RobWiederstein/lasers}
  \item Report bugs at \url{https://github.com/RobWiederstein/lasers/issues}
}

}
\keyword{internal}
