% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_a.R
\name{paste_a}
\alias{paste_a}
\title{Paste A.}
\usage{
paste_a(text)
}
\arguments{
\item{text}{a string}
}
\value{
a character string
}
\description{
Pastes a to front of string
}
\examples{
paste_a("some_string")
}
